requirements.txt
# LLM (DashScope via OpenAI-compatible API)
openai>=1.0.0

# Document loading & chunking
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
unstructured>=0.18.0
markdown>=3.0.0
pypdf>=4.0.0

# Embedding & reranking models
modelscope>=1.14.0
transformers>=4.40.0
torch>=2.2.0

# Vector store
chromadb>=0.5.0

# Utilities
python-dotenv>=1.0.0

# Web UI & visualization
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.0.0
